apps/workers/requirements.txt
celery==5.3.4
redis==5.0.1
nats-py==2.3.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
langchain==0.1.0
langchain-openai==0.0.2
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
reportlab==4.0.4
jinja2==3.1.2
httpx==0.25.2
structlog==23.2.0
python-dotenv==1.0.0
prometheus-client==0.19.0
